requirements.txt
Pillow==10.1.0
NumPy==1.24.3
opencv-python==4.8.0.74
scikit-image==0.21.0